Moved all methods related to configuration-parsing to jirautils.go
